+++ resources/views/back/dashboard.blade.php
@@ -19,13 +19,13 @@
 
                 @if(!$manager->can('index')) @continue @endif
 
-                @if($manager->findAllManaged()->count() > 0)
+                @if($manager->indexCollection()->total() > 0)
                     <div class="column-6">
                         <div class="rounded bg-white shadow">
                             <div class="inset">
                                 <div class="flex items-center mb-4">
-                                    <span class="text-4xl font-bold inline-block mr-4" style="leading-none">{{ $manager->findAllManaged()->count() }}</span>
-                                    <p class="mt-2 text-lg">{{ $manager->findAllManaged()->count() == 1 ? $manager->details()->singular : $manager->details()->plural }}</p>
+                                    <span class="text-4xl font-bold inline-block mr-4" style="leading-none">{{ $manager->indexCollection()->total() }}</span>
+                                    <p class="mt-2 text-lg">{{ $manager->indexCollection()->total() == 1 ? $manager->details()->singular : $manager->details()->plural }}</p>
                                 </div>
                                 <a class="btn btn-primary" href="{{ $manager->route('index') }}">Ga naar {{ $manager->details()->plural }}</a>
                             </div>